added dashboard UI
